impl: remove CI
It's not working (unclear why).

However, it probably *can't* work, as the build requires access to KSP dlls which are not checked in with the repository (for obvious reasons). For a proper CI, I'll need to refactor the mod into a core simulator project which doesn't depend on KSP, and an actual mod project which does. Then the core project can be tested on CI independently of the game files.

This of course would require introducing a layer of abstraction whenever the core project needs to interact with game objects, which will be tricky to implement (especially performance-wise).