impl: major refactoring of vessel shape setup
Here we completely refactor how spacecraft and parts work, with the ability to
construct spacecraft from the `Vessel` tree in an idempotent way. This is
critical for keeping the spacecraft tree in sync as the `Vessel` is loaded and
unloaded through gameplay.

The electrical simulation is effectively disabled for now and will need to be
rewritten on top of the new shape of `SpacecraftComponent`s.