impl: beginnings of the flow-based simulation engine
This implements the `ResourceFlowSimulator`, which simulates resources moving
from producers to consumers by calculating the steady states of those flows
based on available supply and demand. Flows are recalculated whenever the
available production or desired consumption rate of a resource changes.

This is still a very basic system and is largely untested.