impl: major refactoring of electrical simulation
This is a big refactoring of the electrical simulation on top of the new
multithreaded simulation system. It's still not tested in-game, but unit testing
at least shows the new system is capable of simulating the interaction between a
battery and an RTG.

Next step: actual in-game simulation of a battery and RTG working together.