impl: remove thread from `SimulationDriver`
Simplify the `SimulationDriver` significantly by removing its usage of a
`Thread` and a queue of messages. This ended up being unnecessary, as the need
for the game loop to read updated state meant that it would always be blocking
while waiting for state updates anyway.

Instead, the `SimulationDriver` now performs the actual calculations in stable
states in a single-threaded manner. Stabilization (recalculating flow rates for
resources) is by far a more expensive operation and is still performed on the
thread pool.

This ensures that the simulation still runs its main workload in a
multi-threaded manner, while removing overhead for the cheap linear projection.