impl: fix in-game and add an HTTP-based RPC endpoint
This commit fixes the in-game visualization of state, which was broken by the
changes to the way simulation works (the actual simulation was running fine,
just the update of the KSP UI was broken).

Additionally, tihs commit adds an HTTP server to the game mod. This server will
power a companion web UI that serves a few purposes:

* during development, it will allow for better debugging of the mod internals,
  as it can visualize and manipulate state in much more detail than in-game.

* longer term, the main UI of the mod will need to work this way until I either
  develop proper Unity GUI skills or someone builds a real in-game UI :)