impl: overhaul of the `VirtualVessel` system & component init
A large rewrite of the virtual component system, which are the emulated pieces
detached from KSP's `PartModule` logic. Now, `VirtualComponent`s are initialized
when a ship is constructed in the editor, and the editor-configured state of
the component is saved/loaded correctly through their corresponding `PartModule`
(the `HgPartBase` module). Then, when a `Vessel` is created out of those parts,
the components are transitioned over to the `VirtualVessel` side, where they'll
exist and simulate outside of the `PartModule` system.

With this rewrite, the creation, saving, loading, and reloading of ships should
now be more or less correct! Some observations:

* KSP loves to destroy and recreate `VesselModule`s when changing scenes. This
  will lead to some more expensive teardown and re-initialization of
  `VirtualVessel`s. We should implement some caching logic to avoid this.

* The approach to testing used (an `Adapter` interface that keeps KSP-aware and
  non-KSP-aware logic separate) is hopelessly inadequate as the complexity of
  the system here grows. It needs to be replaced with a new approach to testing,
  which will be the next refactoring.