impl: beginnings of the fuel system
This is a rough sketch of how a fuel tank might work. The basic idea is that a
fuel tank part defines a volume, which can be split into one or more tanks, each
containing a different stored material (fuel, oxidizer, etc).

With this commit, tanks are initialized (with LF/Ox) and a basic display is
wired up.

The next step will be wiring up an engine module which is capable of burning
this fuel. That requires figuring out the complexities of resource flow.