refactor: move `HgVirtualPartModule` logic into `VirtualPart` directly
Previously, `HgVirtualPartModule` contained the logic to initialize and link
`VirtualPart`s with the `PartModule` which manages them. However, this structure
has a problem: it's not always feasible to derive `PartModule`s from that base
class. Sometimes, we want to use other KSP `PartModule`s such as `ModuleEngines`
as the base for our parts.

In this refactor, a new interface `IVirtualPartModule` abstracts the contract
between a `VirtualPart` and its `PartModule`, and allows for the logic which
creates and updates a `VirtualPart` to be moved into `VirtualPart` itself (where
it arguably belongs anyway).