impl: simple engine producing thrust
Implements a very simple engine that, when started, generates thrust in
proportion to the throttle. This is incredibly basic, but will serve as a
building block to start adding in more layers of engine functionality.

Hilariously, surface effects on the engine model do seem to work (it glows red)
though I don't fully understand why.