impl: initial pass of the `BufferedResourceConsumer`
This is a utility which manages resource buffering for an engine (or engine-like
entity). It maintains a local buffer of required resources for the engine to run
at full throttle for 1 second of game time, fed from the resource simulation.

This buffering is important as the rate at which a resource's consumption can
change is limited by the simulation frequency, which is capped at 1Hz. If the
engine is throttled low, and the player suddenly sets the throttle to full, we
would otherwise need to wait for the resource simulation to tick to know if the
new rate of resource consumption is valid or not. The `BufferedResourceConsumer`
allows the new demand to be supplied instantly from the buffer, and the buffer
filled in the background from the resource system. This is a bit of a complex
juggling act (and could probably be simplified with some better APIs).

This class is quite untested, but a basic test suite exists at least for the
initialization.