impl: more progress on BufferedResourceConsumer, SimulationDriver refactorings
This commit actually encompasses some larger changes. To begin with,
`SimulationDriver` is refactored to work in terms of integer seconds, rather
than fractional time. This reflects a larger commitment to the concept that
the simulation has a lower bound of 1 Hz. Using integer time simplifies the
logic around resource management. Concurrently with this change, testing
infrastructure is added to make testing `BufferedResourceConsumer` much more
ergonomic and accurate.