impl: adjust fuel tank volumes
HGS uses Liters to measure tank volume. KSP also uses Liters, but because KSP
has unphysical LF/LOX densities and mixing ratio, the total energy contained in
a tank of equal volume is nowhere near equivalent. For that reason, HGS scales
the tank volume by a factor of 5.2, which gives an equivalent energy with KSP's
tanks. That is, a 100U vanilla KSP tank has 520L in HGS.